Move docker plugin into external file
